refactor: split models
