+++ lnbits/core/models/tinyurl.py
@@ -1,0 +1,9 @@
+from pydantic import BaseModel
+
+
+class TinyURL(BaseModel):
+    id: str
+    url: str
+    endless: bool
+    wallet: str
+    time: float
